factor out tnc
