+++ src/components/donation/index.ts
@@ -1,2 +1,3 @@
 export { Steps } from "./Steps";
 export { default as Share } from "./Steps/Share";
+export { default as DonationTermsAndConditions } from "./DonationTermsAndConditions";
