+++ src/pages/Donate/Content.tsx
@@ -3,9 +3,10 @@ import ExtLink from "components/ExtLink";
 import { DappLogo } from "components/Image";
 import QueryLoader from "components/QueryLoader";
 import { Steps } from "components/donation";
-import { APP_NAME, INTERCOM_HELP } from "constants/env";
+import { DonationTermsAndConditions } from "components/donation";
+import { INTERCOM_HELP } from "constants/env";
 import { appRoutes } from "constants/routes";
-import { PRIVACY_POLICY, TERMS_OF_USE_DONOR } from "constants/urls";
+import { PRIVACY_POLICY } from "constants/urls";
 import { memo, useEffect, useRef } from "react";
 import { Link, useLocation } from "react-router-dom";
 import { useIntentQuery } from "services/apes";
@@ -106,17 +107,10 @@ function LoadedContent(props: Props & { intent?: DonationIntent }) {
           />
         </div>
         <FAQ classes="max-md:px-4 md:col-start-2 md:row-span-5 md:w-[18.875rem]" />
-        <p className="max-md:border-t max-md:border-gray-l3 max-md:px-4 max-md:pt-4 col-start-1 text-sm leading-normal text-left text-navy-l1 dark:text-navy-l2">
-          By making a donation to {APP_NAME}, you agree to our{" "}
-          <A href={TERMS_OF_USE_DONOR}>Terms of Service</A>,{" "}
-          <A href={PRIVACY_POLICY}>Privacy Policy</A>. 100% of your donation is
-          tax-deductible to the extent allowed by US law. Your donation is made
-          to {APP_NAME}, a tax-exempt US 501(c)(3) charity that grants
-          unrestricted funds to {props.name} on your behalf. As a legal matter,{" "}
-          {APP_NAME} must provide any donations to {props.name} on an
-          unrestricted basis, regardless of any designations or restrictions
-          made by you. <A href={TERMS_OF_USE_DONOR}>See Terms.</A>
-        </p>
+        <DonationTermsAndConditions
+          endowName={props.name}
+          classes="max-md:border-t max-md:border-gray-l3 max-md:px-4 max-md:pt-4 col-start-1"
+        />
         <p className="max-md:px-4 mb-4 max-mbcol-start-1 text-sm leading-normal text-left text-navy-l1 dark:text-navy-l2">
           <span className="block mb-0.5">
             Need help? See{" "}
